Tohanasa/tohanasa/cv/serializers.py
# tohanasa/cv/serializers.py
from rest_framework import serializers
from .models import CV

class CVSerializer(serializers.ModelSerializer):
    class Meta:
        model = CV
        fields = '__all__'  # Inclut tous les champs du modèle
