requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
whitenoise>=6.6
Pillow>=10.0
requests>=2.31
reportlab>=4.1
pdfkit>=1.0.0
google-generativeai>=0.8
